@@ -3,8 +3,8 @@
   "private": true,
   "author": "Ultimate Gadget Laboratories",
   "main": "electron/dist/electron-main.js",
-  "version": "1.5.14",
-  "firmwareVersion": "8.10.9",
+  "version": "1.5.17",
+  "firmwareVersion": "8.10.12",
   "deviceProtocolVersion": "4.7.0",
   "userConfigVersion": "4.1.1",
   "hardwareConfigVersion": "1.0.0",
@@ -15,11 +15,10 @@
   },
   "license": "See in LICENSE",
   "engines": {
-    "node": ">=14.16.0 <15.0.0",
-    "npm": ">=v6.14.11 <7.0.0"
+    "node": ">=16.13.0 <17.0.0",
+    "npm": ">=8.1.0 <9.0.0"
   },
   "devDependencies": {
-    "jasmine-ts": "0.4.0",
     "@types/command-line-args": "5.0.0",
     "@types/command-line-usage": "5.0.1",
     "@types/copy-webpack-plugin": "8.0.0",
@@ -28,58 +27,56 @@
     "@types/electron-settings": "3.0.0",
     "@types/file-saver": "2.0.1",
     "@types/fs-extra": "9.0.11",
-    "@types/jasmine": "3.6.2",
-    "@types/jasminewd2": "2.0.8",
+    "@types/jasmine": "3.10.0",
     "@types/lodash": "4.14.168",
-    "@types/node": "14.14.37",
+    "@types/node": "16.11.6",
     "@types/node-hid": "1.3.0",
     "@types/request": "2.0.8",
-    "@types/semver": "7.3.4",
+    "@types/semver": "7.3.9",
     "@types/tmp": "0.2.0",
     "@types/yargs": "16.0.1",
+    "@typescript-eslint/eslint-plugin": "^5.9.0",
+    "@typescript-eslint/parser": "^5.9.0",
     "check-node-version": "4.1.0",
     "copy-webpack-plugin": "8.1.1",
     "copyfiles": "2.4.1",
-    "core-js": "3.8.1",
+    "core-js": "3.15.2",
     "cross-env": "7.0.3",
     "decompress": "4.2.1",
     "decompress-targz": "^4.1.1",
     "devtron": "1.4.0",
-    "electron": "12.0.12",
-    "electron-builder": "22.9.1",
+    "electron": "16.0.6",
+    "electron-builder": "22.14.5",
     "electron-debug": "1.5.0",
     "electron-devtools-installer": "2.2.3",
-    "electron-log": "4.2.2",
-    "electron-notarize": "1.0.0",
-    "electron-settings": "3.1.4",
-    "electron-updater": "4.2.5",
+    "electron-log": "4.4.1",
+    "electron-notarize": "1.1.1",
+    "electron-settings": "4.0.2",
+    "electron-updater": "4.6.1",
+    "eslint": "8.6.0",
     "fs-extra": "10.0.0",
-    "jasmine": "3.6.3",
-    "jasmine-core": "3.6.0",
+    "jasmine": "3.10.0",
+    "jasmine-core": "3.10.0",
+    "jasmine-ts": "0.4.0",
     "lerna": "4.0.0",
     "lodash": "4.17.21",
     "node-hid": "2.1.1",
     "npm-run-all": "4.1.5",
     "nrf-intel-hex": "1.3.0",
-    "pre-commit": "1.2.2",
+    "postcss-scss": "4.0.2",
     "request": "2.88.0",
     "rimraf": "3.0.2",
-    "source-map-support": "0.5.9",
-    "stylelint": "13.13.1",
-    "svg-sprite": "1.5.0",
-    "ts-loader": "8.0.17",
-    "ts-node": "9.1.1",
-    "tslint": "6.1.3",
-    "typescript": "4.1.3",
-    "webpack": "5.11.0",
-    "webpack-cli": "4.5.0"
+    "source-map-support": "0.5.20",
+    "stylelint": "14.0.1",
+    "svg-sprite": "1.5.3",
+    "ts-loader": "9.2.6",
+    "ts-node": "10.4.0",
+    "typescript": "4.4.4",
+    "webpack": "5.65.0",
+    "webpack-cli": "4.9.1"
   },
-  "pre-commit": [
-    "precommit-msg"
-  ],
   "scripts": {
     "postinstall": "lerna bootstrap",
-    "precommit-msg": "Git precommit hook is running... & exit 0",
     "test": "lerna run test",
     "lint": "lerna run lint",
     "e2e": "lerna run e2e --scope uhk-web",
