@@ -8,74 +8,80 @@
     "build": "run-s -sn build:renderer",
     "build:web": "ng build --prod --aot",
     "build:renderer": "ng build --prod --aot --project=uhk-renderer",
-    "server:renderer": "ng build --project=uhk-renderer --watch --source-map --extract-css",
+    "server:renderer": "ng build --project=uhk-renderer --watch",
     "test-skip": "ng test --watch false",
     "lint": "run-p -snc lint:*",
     "lint:ng": "ng lint",
-    "lint:renderer": "tslint --project src/tsconfig.renderer.json",
-    "lint:style": "stylelint \"src/**/*.scss\" --syntax scss",
+    "lint:style": "stylelint \"src/**/*.scss\"",
     "pree2e": "webdriver-manager update --standalone false --gecko false --quiet",
     "e2e": "ng e2e"
   },
   "private": true,
   "license": "See in LICENSE",
   "devDependencies": {
-    "@angular/animations": "11.2.4",
-    "@angular-builders/custom-webpack": "11.0.0",
-    "@angular/cdk": "11.2.13",
-    "@angular/cli": "11.2.3",
-    "@angular/common": "11.2.4",
-    "@angular/compiler": "11.2.4",
-    "@angular/compiler-cli": "11.2.4",
-    "@angular/core": "11.2.4",
-    "@angular/localize": "11.2.4",
-    "@angular-devkit/build-angular": "0.1102.3",
-    "@angular-devkit/build-optimizer": "0.1102.3",
-    "@angular-devkit/core": "11.2.3",
-    "@angular/forms": "11.2.4",
-    "@angular/language-service": "11.2.4",
-    "@angular/platform-browser": "11.2.4",
-    "@angular/platform-browser-dynamic": "11.2.4",
-    "@angular/router": "11.2.4",
-    "@ng-bootstrap/ng-bootstrap": "9.1.0",
-    "@ngtools/webpack": "11.2.3",
-    "@ngrx/effects": "10.0.0",
-    "@ngrx/router-store": "10.0.0",
-    "@ngrx/store": "10.0.0",
-    "@ngrx/store-devtools": "10.0.0",
-    "ngx-contextmenu": "5.4.0",
+    "@angular-builders/custom-webpack": "13.0.0",
+    "@angular-devkit/build-angular": " 13.1.3",
+    "@angular-devkit/build-optimizer": "0.1301.3",
+    "@angular-devkit/core": "13.1.3",
+    "@angular-eslint/builder": "13.0.1",
+    "@angular-eslint/eslint-plugin": "13.0.1",
+    "@angular-eslint/eslint-plugin-template": "13.0.1",
+    "@angular-eslint/schematics": "13.0.1",
+    "@angular-eslint/template-parser": "13.0.1",
+    "@angular/animations": "13.1.2",
+    "@angular/cdk": "13.1.2",
+    "@angular/cli": "13.1.3",
+    "@angular/common": "13.1.2",
+    "@angular/compiler": "13.1.2",
+    "@angular/compiler-cli": "13.1.2",
+    "@angular/core": "13.1.2",
+    "@angular/forms": "13.1.2",
+    "@angular/language-service": "13.1.2",
+    "@angular/localize": "13.1.2",
+    "@angular/platform-browser": "13.1.2",
+    "@angular/platform-browser-dynamic": "13.1.2",
+    "@angular/router": "13.1.2",
+    "@fortawesome/angular-fontawesome": "0.7.0",
+    "@fortawesome/fontawesome-svg-core": "1.2.30",
+    "@fortawesome/free-regular-svg-icons": "5.14.0",
+    "@fortawesome/free-solid-svg-icons": "5.14.0",
+    "@materia-ui/ngx-monaco-editor": "6.0.0-beta.1",
+    "@ng-bootstrap/ng-bootstrap": "11.0.0",
+    "@ng-select/ng-select": "8.1.1",
+    "@ngrx/effects": "13.0.2",
+    "@ngrx/router-store": "13.0.2",
+    "@ngrx/store": "13.0.2",
+    "@ngrx/store-devtools": "13.0.2",
+    "@ngtools/webpack": "13.0.3",
     "angular-confirmation-popover": "6.0.0",
-    "angular-notifier": "6.0.1",
+    "angular-notifier": "9.1.0",
     "bootstrap": "4.6.0",
     "codelyzer": "6.0.0",
     "dragula": "3.7.2",
-    "@fortawesome/angular-fontawesome": "0.7.0",
-    "@fortawesome/fontawesome-svg-core": "1.2.30",
-    "@fortawesome/free-solid-svg-icons": "5.14.0",
-    "@fortawesome/free-regular-svg-icons": "5.14.0",
     "jasmine-core": "3.6.0",
     "jasmine-spec-reporter": "5.0.2",
-    "karma": "5.1.0",
+    "karma": "6.3.4",
     "karma-chrome-launcher": "3.1.0",
-    "karma-coverage-istanbul-reporter": "3.0.3",
-    "karma-jasmine": "3.3.1",
-    "karma-jasmine-html-reporter": "1.5.4",
+    "karma-coverage": "2.1.0",
+    "karma-jasmine": "4.0.1",
+    "karma-jasmine-html-reporter": "1.7.0",
+    "monaco-editor": "0.30.1",
     "ng2-dragula": "2.1.1",
     "ng2-nouislider": "1.8.2",
-    "ngx-clipboard": "13.0.1",
-    "ngx-select-ex": "4.0.0",
     "ngrx-store-freeze": "0.2.4",
+    "ngx-clipboard": "14.0.1",
+    "ngx-contextmenu": "6.0.0",
+    "ngx-select-ex": "6.1.0",
     "nouislider": "14.6.0",
     "protractor": "7.0.0",
     "rxjs": "6.6.0",
-    "semver": "7.3.4",
+    "semver": "7.3.5",
     "ts-keycode-enum": "^1.0.6",
     "uhk-common": "1.0.0",
     "xml-loader": "1.2.1",
-    "zone.js": "0.10.3"
+    "zone.js": "0.11.4"
   },
   "dependencies": {
-    "classlist.js": "1.1.20150312",
     "file-saver": "2.0.2",
     "spacing-bootstrap-3": "^1.0.0"
   }
