@@ -78,6 +78,6 @@ export class MacroKeyTabComponent extends MacroBaseComponent implements OnInit {
     isMacroValid = () => {
         const keyMacroAction = this.getKeyMacroAction();
         return !!keyMacroAction.scancode || !!keyMacroAction.modifierMask;
-    }
+    };
 
 }
