@@ -1,25 +1,31 @@
 import {
     AfterViewChecked,
+    ChangeDetectionStrategy,
     Component,
     EventEmitter,
+    forwardRef,
     Input,
+    OnChanges,
+    OnDestroy,
     Output,
     QueryList,
-    ViewChildren,
-    forwardRef,
-    OnDestroy,
-    OnChanges,
-    SimpleChanges
+    SimpleChanges,
+    ViewChildren
 } from '@angular/core';
 import { animate, keyframes, state, style, transition, trigger } from '@angular/animations';
 import { DragulaService } from 'ng2-dragula';
 import { faPlus } from '@fortawesome/free-solid-svg-icons';
 
-import { Macro, MacroAction, KeyMacroAction, KeystrokeAction, MacroKeySubAction } from 'uhk-common';
+import { KeyMacroAction, KeystrokeAction, Macro, MacroAction, MacroKeySubAction } from 'uhk-common';
 
 import { MacroItemComponent } from '../item';
 import { mapLeftRightModifierToKeyActionModifier } from '../../../util';
 import { KeyCaptureData } from '../../../models/svg-key-events';
+
+const ANIMATION_TIME = 500;
+const ANIMATION_INTERVAL = 5;
+const ANIMATION_TIMEOUT = ANIMATION_TIME + ANIMATION_INTERVAL;
+const CANCEL_ACTION_ANIMATION_TIMEOUT = ANIMATION_TIME + 25;
 
 @Component({
     animations: [
@@ -31,12 +37,12 @@ import { KeyCaptureData } from '../../../models/svg-key-events';
                 height: '*'
             })),
             transition('inactive => active',
-                animate('500ms ease-out', keyframes([
+                animate(`${ANIMATION_TIME}ms ease-out`, keyframes([
                     style({ visibility: 'visible', offset: 1 })
                 ]))
             ),
             transition('active => inactive',
-                animate('500ms ease-out', keyframes([
+                animate(`${ANIMATION_TIME}ms ease-out`, keyframes([
                     style({ visibility: 'hidden', offset: 0 })
                 ]))
             )
@@ -44,17 +50,18 @@ import { KeyCaptureData } from '../../../models/svg-key-events';
         trigger('togglerNew', [
             transition(':enter', [
                 style({ height: 0 }),
-                animate('500ms ease-out', style({ height: '*' })
+                animate(`${ANIMATION_TIME}ms ease-out`, style({ height: '*' })
                 )
             ]),
             transition(':leave', [
                 style({ height: '*' }),
-                animate('500ms ease-out', style({ height: 0 })
+                animate(`${ANIMATION_TIME}ms ease-out`, style({ height: 0 })
                 )
             ])
         ])
     ],
     selector: 'macro-list',
+    changeDetection: ChangeDetectionStrategy.OnPush,
     templateUrl: './macro-list.component.html',
     styleUrls: ['./macro-list.component.scss']
 })
@@ -74,6 +81,10 @@ export class MacroListComponent implements AfterViewChecked, OnChanges, OnDestro
     faPlus = faPlus;
     activeEdit: number = undefined;
     scrollTopPosition: number;
+    isMacroReordering = false;
+
+    private scrollToBottomIntervalTimer: number;
+    private scrollToBottomSetTimeoutTimer: number;
 
     constructor(private dragulaService: DragulaService) {
         dragulaService.createGroup(this.MACRO_ACTIONS, {
@@ -100,6 +111,10 @@ export class MacroListComponent implements AfterViewChecked, OnChanges, OnDestro
             window.scrollTo(window.scrollX, this.scrollTopPosition);
             this.scrollTopPosition = undefined;
         }
+
+        if (this.isMacroReordering) {
+            this.isMacroReordering = false;
+        }
     }
 
     ngOnChanges(changes: SimpleChanges): void {
@@ -110,6 +125,12 @@ export class MacroListComponent implements AfterViewChecked, OnChanges, OnDestro
 
     ngOnDestroy(): void {
         this.dragulaService.destroy(this.MACRO_ACTIONS);
+
+        this.clearScrollToBottomInterval();
+
+        if (this.scrollToBottomSetTimeoutTimer) {
+            window.clearTimeout(this.scrollToBottomSetTimeoutTimer);
+        }
     }
 
     showNewAction() {
@@ -117,10 +138,12 @@ export class MacroListComponent implements AfterViewChecked, OnChanges, OnDestro
 
         this.newMacro = undefined;
         this.showNew = true;
+        this.scrollToBottom();
     }
 
     hideNewAction() {
         this.showNew = false;
+        window.setTimeout(() => window.scrollTo(document.body.scrollLeft, document.body.scrollHeight), CANCEL_ACTION_ANIMATION_TIMEOUT);
     }
 
     addNewAction(macroAction: MacroAction) {
@@ -175,6 +198,7 @@ export class MacroListComponent implements AfterViewChecked, OnChanges, OnDestro
     }
 
     macroActionReordered(macroActions: MacroAction[]): void {
+        this.isMacroReordering = true;
         this.scrollTopPosition = window.scrollY;
         this.reorder.emit({
             macroId: this.macro.id,
@@ -183,7 +207,11 @@ export class MacroListComponent implements AfterViewChecked, OnChanges, OnDestro
     }
 
     macroActionTrackByFn(index: number, macroAction: MacroAction): string {
-        return index.toString() + macroAction.toString();
+        if (this.isMacroReordering) {
+            return index.toString() + macroAction.toString();
+        }
+
+        return index.toString();
     }
 
     private toKeyAction(event: KeyCaptureData): KeystrokeAction {
@@ -199,6 +227,21 @@ export class MacroListComponent implements AfterViewChecked, OnChanges, OnDestro
         if (this.activeEdit !== undefined) {
             this.macroItems.toArray()[this.activeEdit].cancelEdit();
             this.activeEdit = undefined;
+        }
+    }
+
+    private scrollToBottom(): void {
+        this.scrollToBottomIntervalTimer = window.setInterval(() => {
+            console.log('scroll', document.body.scrollHeight);
+            window.scrollTo(document.body.scrollLeft, document.body.scrollHeight);
+        }, ANIMATION_INTERVAL);
+
+        this.scrollToBottomSetTimeoutTimer = window.setTimeout(this.clearScrollToBottomInterval.bind(this), ANIMATION_TIMEOUT);
+    }
+
+    private clearScrollToBottomInterval(): void {
+        if (this.scrollToBottomIntervalTimer) {
+            window.clearInterval(this.scrollToBottomIntervalTimer);
         }
     }
 }
