@@ -20,8 +20,7 @@ import { AppState, isForceUpdate } from '../index';
 
 @Injectable()
 export class AppUpdateEffect {
-    @Effect({ dispatch: false })
-    appStart$: Observable<Action> = this.actions$
+    @Effect({ dispatch: false }) appStart$: Observable<Action> = this.actions$
         .pipe(
             ofType(ActionTypes.UpdateApp),
             first(),
