@@ -52,8 +52,7 @@ import { getVersions } from '../../util';
 
 @Injectable()
 export class DeviceEffects {
-    @Effect()
-    deviceConnectionStateChange$: Observable<Action> = this.actions$
+    @Effect() deviceConnectionStateChange$: Observable<Action> = this.actions$
         .pipe(
             ofType<ConnectionStateChangedAction>(ActionTypes.ConnectionStateChanged),
             withLatestFrom(this.store.select(getRouterState), this.store.select(deviceConnected)),
@@ -121,8 +120,7 @@ export class DeviceEffects {
             })
         );
 
-    @Effect({ dispatch: false })
-    setPrivilegeOnLinux$: Observable<Action> = this.actions$
+    @Effect({ dispatch: false }) setPrivilegeOnLinux$: Observable<Action> = this.actions$
         .pipe(
             ofType(ActionTypes.SetPrivilegeOnLinux),
             tap(() => {
@@ -130,8 +128,7 @@ export class DeviceEffects {
             })
         );
 
-    @Effect()
-    setPrivilegeOnLinuxReply$: Observable<Action> = this.actions$
+    @Effect() setPrivilegeOnLinuxReply$: Observable<Action> = this.actions$
         .pipe(
             ofType<SetPrivilegeOnLinuxReplyAction>(ActionTypes.SetPrivilegeOnLinuxReply),
             map(action => action.payload),
@@ -145,8 +142,7 @@ export class DeviceEffects {
             })
         );
 
-    @Effect({ dispatch: false })
-    saveConfiguration$: Observable<Action> = this.actions$
+    @Effect({ dispatch: false }) saveConfiguration$: Observable<Action> = this.actions$
         .pipe(
             ofType<SaveConfigurationAction>(ActionTypes.SaveConfiguration),
             withLatestFrom(this.store),
@@ -160,8 +156,7 @@ export class DeviceEffects {
             switchMap(() => EMPTY)
         );
 
-    @Effect()
-    saveConfigurationReply$: Observable<Action> = this.actions$
+    @Effect() saveConfigurationReply$: Observable<Action> = this.actions$
         .pipe(
             ofType<SaveConfigurationReplyAction>(ActionTypes.SaveConfigurationReply),
             map(action => action.payload),
@@ -182,8 +177,7 @@ export class DeviceEffects {
             })
         );
 
-    @Effect()
-    autoHideSaveToKeyboardButton$: Observable<Action> = this.actions$
+    @Effect() autoHideSaveToKeyboardButton$: Observable<Action> = this.actions$
         .pipe(
             ofType(ActionTypes.SaveToKeyboardSuccess),
             withLatestFrom(this.store),
@@ -203,8 +197,7 @@ export class DeviceEffects {
             )
         );
 
-    @Effect()
-    resetMouseSpeedSettings$: Observable<Action> = this.actions$
+    @Effect() resetMouseSpeedSettings$: Observable<Action> = this.actions$
         .pipe(
             ofType(
                 ActionTypes.ResetPcMouseSpeedSettings,
@@ -227,22 +220,22 @@ export class DeviceEffects {
         .pipe(
             ofType<LoadResetUserConfigurationAction>(UserConfigActions.LoadResetUserConfiguration),
             map(action => action.payload),
-            switchMap((config: UserConfiguration) => {
-                this.dataStorageRepository.saveConfig(config);
-
-                return of(new SaveConfigurationAction(true));
-            })
+            switchMap((config: UserConfiguration) => this.dataStorageRepository.saveConfig(config)
+                .pipe(
+                    map(() => new SaveConfigurationAction(true))
+                )
+            )
         );
 
     @Effect() applyUserConfigurationFromFileAction$ = this.actions$
         .pipe(
             ofType<ApplyUserConfigurationFromFileAction>(UserConfigActions.ApplyUserConfigurationFromFile),
             map(action => action.payload),
-            switchMap(payload => {
-                this.dataStorageRepository.saveConfig(payload.userConfig);
-
-                return of(new SaveConfigurationAction(payload.saveInHistory));
-            })
+            switchMap(payload => this.dataStorageRepository.saveConfig(payload.userConfig)
+                .pipe(
+                    map(() => new SaveConfigurationAction(payload.saveInHistory))
+                )
+            )
         );
 
     @Effect({ dispatch: false }) updateFirmware$ = this.actions$
