@@ -8,7 +8,6 @@ import {
     createMd5Hash,
     getMd5HashFromFilename,
     HardwareModules,
-    Keymap,
     LEFT_HALF_MODULE,
     PlayMacroAction,
     UHK_60_DEVICE,
@@ -16,8 +15,8 @@ import {
     UserConfiguration
 } from 'uhk-common';
 
+import * as fromDefaultUserConfig from './reducers/default-user-configuration.reducer';
 import * as fromUserConfig from './reducers/user-configuration';
-import * as fromPreset from './reducers/preset';
 import * as fromAppUpdate from './reducers/app-update.reducer';
 import * as fromContributors from './reducers/contributors.reducer';
 import * as autoUpdateSettings from './reducers/auto-update-settings';
@@ -28,7 +27,7 @@ import * as fromUserConfigHistory from './reducers/user-configuration-history.re
 import * as fromSelectors from './reducers/selectors';
 import { initProgressButtonState } from './reducers/progress-button-state';
 import { environment } from '../../environments/environment';
-import { RouterStateUrl } from './router-util';
+import { RouterState } from './router-util';
 import { PrivilagePageSate } from '../models/privilage-page-sate';
 import { isVersionGte } from '../util';
 import {
@@ -45,11 +44,11 @@ import { SelectOptionData } from '../models/select-option-data';
 
 // State interface for the application
 export interface AppState {
+    defaultUserConfiguration: fromDefaultUserConfig.State;
     userConfiguration: fromUserConfig.State;
-    presetKeymaps: Keymap[];
     autoUpdateSettings: autoUpdateSettings.State;
     app: fromApp.State;
-    router: RouterReducerState<RouterStateUrl>;
+    router: RouterReducerState<RouterState>;
     appUpdate: fromAppUpdate.State;
     device: fromDevice.State;
     contributors: fromContributors.State;
@@ -58,8 +57,8 @@ export interface AppState {
 }
 
 export const reducers: ActionReducerMap<AppState> = {
+    defaultUserConfiguration: fromDefaultUserConfig.reducer,
     userConfiguration: fromUserConfig.reducer,
-    presetKeymaps: fromPreset.reducer,
     autoUpdateSettings: autoUpdateSettings.reducer,
     app: fromApp.reducer,
     router: routerReducer,
@@ -148,29 +147,29 @@ export const flashFirmwareButtonDisabled = createSelector(runningInElectron, upd
 export const getStateHardwareModules = createSelector(deviceState, fromDevice.getHardwareModules);
 export const getHardwareModules = createSelector(runningInElectron, getStateHardwareModules, getAgentVersionInfo,
     (electron, hardwareModules, agentVersionInfo): HardwareModules => {
-    if (electron) {
-        return hardwareModules;
-    }
-
-    return {
-        moduleInfos: [
-            {
-                module: LEFT_HALF_MODULE,
-                info: {
-                    firmwareVersion: agentVersionInfo.firmwareVersion,
-                    moduleProtocolVersion: agentVersionInfo.moduleProtocolVersion
-                }
-            }
-        ],
-        rightModuleInfo: {
-            deviceProtocolVersion: agentVersionInfo.deviceProtocolVersion,
-            hardwareConfigVersion: agentVersionInfo.hardwareConfigVersion,
-            firmwareVersion: agentVersionInfo.firmwareVersion,
-            moduleProtocolVersion: agentVersionInfo.moduleProtocolVersion,
-            userConfigVersion: agentVersionInfo.userConfigVersion
+        if (electron) {
+            return hardwareModules;
         }
-    };
-});
+
+        return {
+            moduleInfos: [
+                {
+                    module: LEFT_HALF_MODULE,
+                    info: {
+                        firmwareVersion: agentVersionInfo.firmwareVersion,
+                        moduleProtocolVersion: agentVersionInfo.moduleProtocolVersion
+                    }
+                }
+            ],
+            rightModuleInfo: {
+                deviceProtocolVersion: agentVersionInfo.deviceProtocolVersion,
+                hardwareConfigVersion: agentVersionInfo.hardwareConfigVersion,
+                firmwareVersion: agentVersionInfo.firmwareVersion,
+                moduleProtocolVersion: agentVersionInfo.moduleProtocolVersion,
+                userConfigVersion: agentVersionInfo.userConfigVersion
+            }
+        };
+    });
 export const getBackupUserConfigurationState = createSelector(deviceState, fromDevice.getBackupUserConfigurationState);
 export const getRestoreUserConfiguration = createSelector(deviceState, fromDevice.getHasBackupUserConfiguration);
 export const bootloaderActive = createSelector(deviceState, fromDevice.bootloaderActive);
@@ -299,12 +298,12 @@ export const getSideMenuPageState = createSelector(
     calculateDeviceUiState,
     getConnectedDevice,
     (showAddonMenuValue: boolean,
-     runningInElectronValue: boolean,
-     updatingFirmwareValue: boolean,
-     userConfiguration: UserConfiguration,
-     restoreUserConfiguration: boolean,
-     uiState,
-     connectedDevice): SideMenuPageState => {
+        runningInElectronValue: boolean,
+        updatingFirmwareValue: boolean,
+        userConfiguration: UserConfiguration,
+        restoreUserConfiguration: boolean,
+        uiState,
+        connectedDevice): SideMenuPageState => {
         return {
             connectedDevice: runningInElectronValue ? connectedDevice : UHK_60_DEVICE,
             showAddonMenu: showAddonMenuValue,
@@ -342,7 +341,7 @@ export const getApplicationSettings = createSelector(
     appUpdateSettingsState,
     appState,
     (updateSettingsState,
-     app
+        app
     ): ApplicationSettings => {
         return {
             checkForUpdateOnStartUp: updateSettingsState.checkForUpdateOnStartUp,
@@ -359,15 +358,32 @@ export const getUserConfigHistoryComponentState = createSelector(
     getMd5HasOfUserConfig,
     isUserConfigSaving,
     (inElectron,
-     state: fromUserConfigHistory.State,
-     md5Hash: string,
-     saving: boolean): UserConfigHistoryComponentState => {
+        state: fromUserConfigHistory.State,
+        md5Hash: string,
+        saving: boolean): UserConfigHistoryComponentState => {
+        let foundFirstCurrent = false;
+
         return {
             loading: inElectron && state.loading,
-            files: state.files.map(x => ({
-                file: x,
-                showRestore: getMd5HashFromFilename(x) !== md5Hash
-            })),
+            files: state.files.map(x => {
+                const showRestore = getMd5HashFromFilename(x) !== md5Hash;
+                let displayText: string;
+
+                if (showRestore) {
+                    displayText = 'Restore';
+                } else if (foundFirstCurrent) {
+                    displayText = 'Same as current';
+                } else {
+                    displayText = 'Current';
+                    foundFirstCurrent = true;
+                }
+
+                return {
+                    displayText,
+                    showRestore,
+                    file: x
+                };
+            }),
             disabled: saving
         };
     });
@@ -409,3 +425,9 @@ export const getFirmwareUpgradeState = createSelector(runningInElectron, getStat
             recoveryModules: []
         };
     });
+
+export const defaultUserConfigState = (state: AppState) => state.defaultUserConfiguration;
+export const getDefaultUserConfigurationKeymaps = createSelector(
+    defaultUserConfigState, fromDefaultUserConfig.getDefaultUserConfigurationKeymaps);
+export const getSelectedAddKeymap = createSelector(
+    defaultUserConfigState, fromDefaultUserConfig.getSelectedKeymap);
