@@ -5,7 +5,7 @@ import { UhkRendererModule } from './renderer/uhk-renderer.module';
 import { environment } from './environments/environment';
 
 if (environment.production) {
-  enableProdMode();
+    enableProdMode();
 }
 
 platformBrowserDynamic().bootstrapModule(UhkRendererModule, {
